✏️ Rename: `newsSlider.svelte` to upper camel case [#54]
