+++ src/lib/components/NewsSlider.svelte
